Minor Text Changes
Feel free to disregard if they don't make sense. 

However, the phrase below must be corrected do to an artifact error. 

"To understand the impact of turning a feature on or off in a Feature" 

to 

"To understand the impact of turning a feature on or off in a environment"