back to react 17 with new plugin update (#300)
Docusaurus plugin required react 18 since our sdks do (react 17 had a bug), but docusaurus doesn't officially support 18 until version 3. So an update to the plugin allows it to work with react 17, meaning this will be easier for users to install without having to bump react to 18 and seeing dev warnings.